add auto_resub flag for each user; check flag is false to unsub user on notification
- added methods to database.py to get/set auto_resub flag (True if user wants to stay subscribed, False if not)
- user is not auto resubscribed by default
- include check in notify.py: if user does not opt to auto resub, then remove them from subscription list immediately

+++ src/notify.py
@@ -102,7 +102,8 @@ class Notify:
                     Client(TWILIO_SID, TWILIO_TOKEN).api.account.messages.create(
                         to=f"+1{phone}", from_=TWILIO_PHONE, body=msg
                     )
-                self.db.remove_from_waitlist(self._netids[i], self._classid)
+                if not self.db.get_auto_resub(self._netids[i]):
+                    self.db.remove_from_waitlist(self._netids[i], self._classid)
             return True
         except Exception as e:
             print(e, file=stderr)
